// src/types/deviceTypes.ts
export type DeviceDBType = {
    id: string
    ip: string
    title: string
    issuedAt: string
    expiredAt: string
    deviceId: string
    userId: string
    isValid: boolean
}

export type DeviceType = {
    ip: string
    title: string
    issuedAt: string
    expiredAt: string
    deviceId: string
    userId: string
    isValid: boolean
}